o Add #ifdef APPEND_ADDRESS, default off, appends function address to function   name when auto-profiled. o Left-justify symbol name for printing in all cases o pr_summary.c: Initialize global, global_p to 0, so that when one of them was   not allocated (i.e. no regions), the test e.g. if (global) free (global)   behaves properly.

+++ tests/summary.c
@@ -121,7 +121,7 @@ double sub (int t)
   usec = 1000 * (iam + t)*10;
 
   ret = GPTLstart ("sleep");
-  printf ("iam %d thread %d sleeping %d ms\n", iam, omp_get_thread_num(), usec/1000);
+  printf ("iam %d thread %d sleeping %d ms\n", iam, omp_get_thread_num(), (int) usec/1000);
   usleep (usec);
   ret = GPTLstop ("sleep");
